requirements.txt
streamlit==1.28.1
gfpgan==1.3.8
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
tqdm==4.66.1
basicsr==1.4.2
facexlib==0.3.0
realesrgan==0.3.0